rename #459: heart -> like
