+++ be/src/test/java/com/foodymoody/be/acceptance/feed_like/FeedLikeAcceptanceTest.java
@@ -1,12 +1,12 @@
 package com.foodymoody.be.acceptance.feed_like;
 
 import static com.foodymoody.be.acceptance.feed.FeedSteps.피드를_등록한다;
-import static com.foodymoody.be.acceptance.feed_like.FeedHeartSteps.응답코드가_200이고_id가_존재하면_정상적으로_좋아요_가능;
-import static com.foodymoody.be.acceptance.feed_like.FeedHeartSteps.응답코드가_204이면_정상적으로_좋아요_취소;
-import static com.foodymoody.be.acceptance.feed_like.FeedHeartSteps.좋아요_취소를_한다;
-import static com.foodymoody.be.acceptance.feed_like.FeedHeartSteps.좋아요_한_적이_없는데_좋아요_취소를_한다;
-import static com.foodymoody.be.acceptance.feed_like.FeedHeartSteps.좋아요된_피드에_또_좋아요를_한다;
-import static com.foodymoody.be.acceptance.feed_like.FeedHeartSteps.좋아요를_한다;
+import static com.foodymoody.be.acceptance.feed_like.FeedLikeSteps.응답코드가_200이고_id가_존재하면_정상적으로_좋아요_가능;
+import static com.foodymoody.be.acceptance.feed_like.FeedLikeSteps.응답코드가_204이면_정상적으로_좋아요_취소;
+import static com.foodymoody.be.acceptance.feed_like.FeedLikeSteps.좋아요_취소를_한다;
+import static com.foodymoody.be.acceptance.feed_like.FeedLikeSteps.좋아요_한_적이_없는데_좋아요_취소를_한다;
+import static com.foodymoody.be.acceptance.feed_like.FeedLikeSteps.좋아요된_피드에_또_좋아요를_한다;
+import static com.foodymoody.be.acceptance.feed_like.FeedLikeSteps.좋아요를_한다;
 import static com.foodymoody.be.acceptance.image.ImageSteps.피드_이미지를_업로드한다;
 
 import com.foodymoody.be.acceptance.AcceptanceTest;
