+++ be/src/test/java/com/foodymoody/be/acceptance/notification/NotificationAcceptanceTest.java
@@ -11,7 +11,7 @@ import static com.foodymoody.be.acceptance.feed_collection_like.FeedCollectionLi
 import static com.foodymoody.be.acceptance.feed_collection_reply.FeedCollectionReplySteps.피드_컬렉션_댓글에_대댓글을_등록하고_아이디를_반환한다;
 import static com.foodymoody.be.acceptance.feed_collection_reply.FeedCollectionReplySteps.피드_컬렉션_댓글에_대댓글을_등록한다;
 import static com.foodymoody.be.acceptance.feed_collection_reply_like.FeedCollectionReplyLikeSteps.피드_컬렉션_대댓글에_좋아요를_등록한다;
-import static com.foodymoody.be.acceptance.feed_like.FeedHeartSteps.좋아요를_한다;
+import static com.foodymoody.be.acceptance.feed_like.FeedLikeSteps.좋아요를_한다;
 import static com.foodymoody.be.acceptance.image.ImageSteps.피드_이미지를_업로드한다;
 import static com.foodymoody.be.acceptance.member.MemberSteps.팔로우한다;
 import static com.foodymoody.be.acceptance.notification.NotificationSteps.모든_알림을_읽음으로_변경한다;
